+++ sources/@roots/create-bud-app/src/tasks/write.src.ts
@@ -6,26 +6,37 @@ export default async function writeSrcTask(command: CreateCommand) {
   const spinner = command.createSpinner()
   spinner.start(`Writing src/**/*...`)
 
-  if (command.support.includes(`react`)) {
-    await command.fs.copy(
-      join(command.createRoot, `templates`, `react`, `src`),
-      `src`,
-    )
-    return spinner.succeed()
+  if (!command.overwrite && command.exists(`src`)) {
+    return spinner.warn(`src already exists. skipping write task.`)
   }
 
-  if (command.support.includes(`vue`)) {
+  try {
+    if (command.support.includes(`react`)) {
+      await command.fs.copy(
+        join(command.createRoot, `templates`, `react`, `src`),
+        `src`,
+        {overwrite: true},
+      )
+      return spinner.succeed()
+    }
+
+    if (command.support.includes(`vue`)) {
+      await command.fs.copy(
+        join(command.createRoot, `templates`, `vue`, `src`),
+        `src`,
+        {overwrite: true},
+      )
+      return spinner.succeed()
+    }
+
     await command.fs.copy(
-      join(command.createRoot, `templates`, `vue`, `src`),
+      join(command.createRoot, `templates`, `default`, `src`),
       `src`,
+      {overwrite: true},
     )
-    return spinner.succeed()
+    spinner.succeed()
+  } catch (error) {
+    spinner.fail()
+    throw error
   }
-
-  await command.fs.copy(
-    join(command.createRoot, `templates`, `default`, `src`),
-    `src`,
-  )
-
-  spinner.succeed()
 }
